Crate parachain_runtime_mock

Structs

AccountIdToMultiLocation

Ancestry

AssetLocation

BaseRate

BaseXcmWeight

BlockHashCount

BlockLength

CurrencyIdConvert

ExistentialDeposit

ExistentialDeposits

GenesisConfig

KsmPerSecond

MaxAssetsForTransfer

MaxInstructions

MaxLocks

MaxReserves

MockBlockNumberProvider

MultiPaymentCurrencySetFee

NativeAssetId

NativeExistentialDeposit

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainMinFee

ParachainNativeCurrencyId

ParachainRuntime

ParachainSpotPriceProviderStub

ParachainTreasuryAccount

RegistryStringLimit

RelayChainOrigin

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

SS58Prefix

SelfLocation

SequentialIdOffset

TransactionByteFee

TreasuryPalletId

UnitWeightCost

VestingPalletId

XcmConfig

Enums

Call

Event

OriginCaller

Constants

CORE_ASSET_ID

MOCK_BLOCK_NUMBER_PROVIDER

NORMAL_DISPATCH_RATIO

Traits

InherentDataExt 🔒

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Amount

AssetRegistry

Balances

BalancesConfig

Barrier

Block 🔒

BlockNumber

CumulusXcm

Currencies

DmpQueue

LocalAssetTransactor

LocalOriginToLocation

LocationToAccountId

MultiTransactionPayment

MultiTransactionPaymentConfig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

SlowAdjustingFeeUpdate

System

SystemConfig

Tokens

TransactionPayment

UncheckedExtrinsic 🔒

UnknownTokens

XTokens

XcmOriginToCallOrigin

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.

XcmpQueue